[Fixed] Loot min/max from int to float
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3/trunk@4305 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ MMOCoreORB/src/server/zone/managers/loot/LootManagerImplementation.cpp
@@ -80,8 +80,8 @@ void LootManagerImplementation::setInitialObjectStats(LootItemTemplate* template
 	if (tanoTemplate != NULL) {
 		Vector<String>* titles = tanoTemplate->getExperimentalGroupTitles();
 		Vector<String>* props = tanoTemplate->getExperimentalSubGroupTitles();
-		Vector<int>* mins = tanoTemplate->getExperimentalMin();
-		Vector<int>* maxs = tanoTemplate->getExperimentalMax();
+		Vector<float>* mins = tanoTemplate->getExperimentalMin();
+		Vector<float>* maxs = tanoTemplate->getExperimentalMax();
 		Vector<short>* prec = tanoTemplate->getExperimentalPrecision();
 
 		for (int i = 0; i < props->size(); ++i) {
